backend/requirements.txt
flask
flask-cors
openai-whisper
ffmpeg-python
torch
requests